crawl/requirements.txt
scrapy
jsonlines
jsonschema
azure-storage-blob
scrapy-impersonate
